Move other components to separate files. Remove conditional tab.
